Login and Sign up pages created
A login page with a form to choose email and password authentication and
an option for google login was implimented.
an option to create an account below the login form is placed and
minipulates the state to show
a sign in option with both email/password and google authentication
methods to provide the user with a choice.