ReactRouter, useContext & Google OAuth
This large commit updates the code with fully functional Good OAuth
to grab and store some user credientials that can be looked for with
a useEffect and updates the context of the user and profile.

React Router DOM was created for navgiating through the application with
ease and created an easy way to show the user a dynamic navigation
depending on weather or not they are logged in.

Styling and animations are still being worked on in the navigation and
the development of the Landing page/Home page will start next. Along with
using the new user credentials to connect, store and fetch data with Sanity.io