Creating animations, constants and UI/UX
A constants directory was created for showing static data in the JSX easily and efficiently.
Inside the directory I created one file for showing demoCards on the home page for each blog post
and each were rendered to the UI.

Motion was imported to start animating the new components created on the home page and a new font was added to
go better with the styling of the app. More of a fun like nature.

My favorite animation creation is for the new navigation functionality where on laptop screen the navigation
will receed itself out of view until a user hovers over the arrow component created for easily accessing the navigation