Dynamically Setting Routes
Pages were created for a logged in user that take the user to
the blogs page, their profile and their dashboard.
Dynamic Routes were created to make sure the user
is not capable of accessing the login/signup page.

For styling a cool new map photo with people profile dummy data was
placed on the homepage.