Connecting with Sanity
In this commit many implementations were put forth, most importantly
a new connection to Sanity was established and a few simple queries that
will be fetching the blog/author/catagories stored in the CMS.

Another important feature was an update to the router and navigation
for the user conditionally on if they are currently logged in or not.

Still deciding on how to continuly authenticate users with refresh tokens
via a server.