Now A User Can View A Specific Post
By creating a route that takes the client to a dynamic url, the client can now view a specific post
that will carry much more information and the actual body of the post to read about.
It will also carry all of the information pertaining to when it was last updated, who wrote the post & when,
a link to that users profile, and much more!