New User Profile Route
Now in this amazing application, via a blog post the client can click on
a username to display a bloggers profile!!

On the user profile page the client will have no CRUD capabilities
besides to Read the data. They will se their blog posts, their
followers(yet to be implememted) their time with the app and muxh more!