Deleting CKEditor
Because of the lack of ease integrating an image picker with CKEditor I choose to instead uninstall their
react package and intergrate a new rich text editor with ease of image uploading & url uploading.

CKEditor with images requires you to create multiple accounts with them and subscriptions if functionality is to
be of a decent standard which defeats the purpose of this blog and how I am writting the code for it.

As my first you tube video, I plan to walk people through creating this website and plan for it to be of the uttmost ease
for them.