Fetching User Profiles
Before covering user profiles, I will mention first that a new text editor was installed with React that is much easier
to implement into the code for this project called Jodit-React. This text editor is now functional in the new post component.

Now, the user profile page has access to a specific user by userId via useParams() hook. This page now displays a user profile
including their picture and name, their time on this website and the posts that they have created.