Adding A Fallback User Photo
Because of occasional errors loading a Google profile photo, the user will now see a user avatar where there photo should be
instead of an alt attribute value.

This is important for the user experience as it repels the user from believing there is an error in the application itself
and instead there is simply an issue loading their photo instead.