Update Your Username
A user can now successfully update their username from the profile page by simply inputing what they want it to be
and submitting the new name.

This is important for the user exprience as it allows for customization.

No error handling has been implemented yet. Soon error handling will be implemented along with the client checking for exsisting user
names in the database so their is no confusion differentiating between users.