So Much Functionality!!!!
More awesome features were added to the app including, saving/liking a post! A user will see how many likes a post has and if they have already liked
a post.

Another feature is the search functionality in the Blog component. A user now can search posts via input on the top of the page.
The search is live and will find posts related to your search term as you type.

The user now can update their username with a safety dynamic confirmation component before submitting their new username request.

These features are a fantastic way to keep people involved and informed about how their blogs are doing, along with finding great blogs to read in an easy way,
weather that is for research, entertainment, education or just for fun.