Handling New Blog Post Functionality
This is a difficult one to figure out guys. the more freedom a client has to create, the more logic and orginization is needed in order to
give the desired result.

A newBlogContext was created to save a users input as long as the user has not refreshed the page. Localstorage will be put to use to store the blog as a draft
to prevent lost blogs when logged out after an hour or an accidental refresh happens.

Now a user can set the header image to the blog post with a file input upload tag.

Much more work is to be done to create a new blog.