More Real Time Liking Attempts
My bug has contended with knowledge on quite a large scale with this one.

Trying different ways to fetch the updated version of the posts right away
after a patch request from the client to update in real time wether or not
you have liked the post, but something is not updating on time with the
getPosts() query.