Sanity: 1, Ryan:
I had my head twisted around this one
for a while, until I realized I had a
special configuration on my Sanity Client.

Sanity provides a CDN to the consumers of
its API. This CDN allows for chachable data to
be sent back to the client for extremly fast
response times. This is normally a
great way to utilize sanity in an app. In
my specific case it is not optimal as this was
stopping me from fetching the most up to date
data when the goal was to show new likes in real time.

useCDN on sanityClient config was set to false.
This fixes the issue of real time data fetching.

A new problem will now persit as the application
grows not that the CDN is not in use. That
being fetching large amounts of data repeatedly in a short period of time, such
as blog posts. I have a few ideas for a solution such
as pagination, or a new client instance.

This will be researched further