Comments, Bio, Relations
What a large commit!! In this commit the comments
section was completed for each post. Now a user
can comment on a blog and a view each comment
made on the specific blog from the blog details page.

Another great feauture on this page is the ability to now link to
other blogs that specific creator has made.

Two last fun pieces of logic added is on the profile page.
now anuser can see the total likes and comments he has
aquired, how many posts he has contributed to the public
and the ability to add and display their bio.