Excepting Dangerously Set InnerHTML
Jodit editor return HTML as the content when using the text editor. This becomes an issue with
rendering the blog content to the page because of the need for setting innerHtML from the server.

A cleansing will be done on the HTML before rendering.

In this commit, the beginning of the functionality for posting a new blog has been implemented. The new post
still needs the configuration of the postedBy materials, excepting catagories, adding a destination and correctly
setting a new image and storying it in sanity's database.