Query By Category
Now, not only does a user see the categories
used for each blog post, but can now query posts
which contain that category in the corrosponding
categories refrencing array.

By clicking a category listed on the blog post
the state which holds the array of posts will be replaced by posts which contain the clicked on category.

This will make it easy for users to find related blog posts straight from the blog page.