useEffect: 1, Ryan: 0
I was yet again bested by our best friend useEffect. It seemed to be that issues continued to occure when trying to update the
new blog with sanity client in a patch request.

The error was occurring when trying to set the categories array in the post. A function was created to check weather or not
an id already exsists in the categories array. If the id did exist the filter method waas called and state set accordingly and
the state was set occorrding to if the id was not present as well.

This function was called by a useEffect when first rendering the component to the DOM each time a category was mapped over pulling from
the localStorage were I was setting the state and populating each editable element.

This is the kicker... In dev enviorment useEffect is called twice on render, successfully calling the add function for each map before setting the
updated state call could complete.

No errors truly existed. useEffect only made it appear to be the case. Rookie me!