Search Errors
Handling a search error!? Me!? Wow, I have come a long ways as a developer since a year ago when I started.
It used to be that I never handled errors because I was scared of them. WHAT FOR!? I have no idea.

Now When a user searches for a post by title/category etc... And no posts show up, the website will automatically rerender
posts so you are not stuck on the loading page wondering why it is taking so long.

The layout to the blog page was also changed, so now the website will only conditionally render blog posts themselves.
More changes in this direction with fixing the layout of the code base will dramatically reduce api calls. More progress will endure