Follow & unfollow
The user can now safely follow other user profiles to see their blog posts.
The bug before introduced mulitple fields in the follow array while the query was being processed where now the
user is not capable of resending the query again before it has finsihed.

Another benefit added to the website is now the user no longer needs to refresh the page in order to update the content. A new query is made behind the scene after the query has completed to
follow or unfollow a user and allows React to do it's magic in an instant.