Logout()
Calling logout was causing unexpected behaviors
within the application. This was due to using the useNavigate hook from react-router-dom.

The user was cleared from all storage and context, but the application would not register the state of the user when navigating back to the dashboard or
or any other user authenticated route.

This causes an unpleasant user experience as
it will cause blank pages, glitches and other side effect that drive away traffic.